requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-cloud-speech==2.24.1
python-multipart==0.0.9
websockets==12.0
python-dotenv==1.0.1 